+++ test/validator/response.test.ts
@@ -418,4 +418,32 @@ describe('Response Validator', () => {
 
 		expect(response).toEqual([200, 418, 422])
 	})
+
+	it('validate nested references', async () => {
+		const job = t.Object(
+			{
+				name: t.String()
+			},
+			{ $id: 'job' }
+		)
+
+		const person = t.Object({
+			name: t.String(),
+			job: t.Ref(job)
+		})
+
+		const app = new Elysia().model({ job, person }).get(
+			'/',
+			() => ({
+				name: 'sucrose',
+				job: { name: 'alchemist' }
+			}),
+			{
+				response: person
+			}
+		)
+
+		const res = await app.handle(req('/'))
+		expect(res.status).toBe(200)
+	})
 })