+++ example/a.ts
@@ -1,16 +1,45 @@
 import { Elysia, t } from '../src'
 
-const app = new Elysia({ precompile: true })
-	.onRequest((ctx) => {
-		console.log(ctx.server?.requestIP(ctx.request))
+const OnlyStrongPassword = new Elysia()
+	.guard({
+		body: t.Object(
+			{ password: t.Literal('strong-password') },
+			{ additionalProperties: true }
+		),
+		beforeHandle() {
+			console.log('Checking password...')
+		}
 	})
-	.get('/', () => 'Hello, World!')
+	.as('plugin')
+
+const app = new Elysia()
+	.use(OnlyStrongPassword)
+	.guard({
+		body: t.Object(
+			{ username: t.Literal('Jhon') },
+			{ additionalProperties: true }
+		),
+		beforeHandle() {
+			console.log('Checking name...')
+		}
+	})
+	.post('/test', ({ body }) => {
+		return 'Hello ' + body.username
+	})
 	.listen(3000)
 
-console.log(`Listening on http://${app.server!.hostname}:${app.server!.port}`)
+const response = await app
+	.handle(
+		new Request('http://localhost/test', {
+			// notice how password is wrong
+			body: JSON.stringify({
+				password: 'weak-password',
+				username: 'Jhon'
+			}),
+			method: 'POST',
+			headers: { 'Content-Type': 'application/json' }
+		})
+	)
+	.then((res) => res.text())
 
-async function test() {
-	const response = await fetch('http://localhost:3000')
-	console.log(await response.text())
-}
-test()
+console.log(response)